+++ changelog.md
@@ -11,6 +11,7 @@
 * Add logs for add/rm/del atom and add/del links
 * [Issue 940](https://github.com/AmpersandTarski/Ampersand/issues/940) Remove company logos from footer
 * [Issue 951](https://github.com/AmpersandTarski/Ampersand/issues/951) Refresh page after role (de)select
+* [Issue 983](https://github.com/AmpersandTarski/Ampersand/issues/983) Increase default timeout for installing the application to 5 min
 
 ## v1.3.0 (15 july 2019)
 * Bugfix error message in case of network/connection error